dbeaver/pro#8126 review fix
