+++ plugins/org.jkiss.dbeaver.registry/src/org/jkiss/dbeaver/registry/BasePlatformImpl.java
@@ -35,6 +35,7 @@ import org.jkiss.dbeaver.model.impl.preferences.AbstractPreferenceStore;
 import org.jkiss.dbeaver.model.navigator.DBNModel;
 import org.jkiss.dbeaver.model.net.DBWHandlerRegistry;
 import org.jkiss.dbeaver.model.preferences.DBPPreferenceStore;
+import org.jkiss.dbeaver.model.runtime.DBRProgressMonitor;
 import org.jkiss.dbeaver.model.runtime.OSDescriptor;
 import org.jkiss.dbeaver.model.sql.SQLDialectMetadataRegistry;
 import org.jkiss.dbeaver.model.task.DBTTaskController;
@@ -46,13 +47,19 @@ import org.jkiss.dbeaver.registry.network.NetworkHandlerRegistry;
 import org.jkiss.dbeaver.registry.settings.GlobalSettings;
 import org.jkiss.dbeaver.runtime.IPluginService;
 import org.jkiss.dbeaver.runtime.jobs.DataSourceMonitorJob;
+import org.jkiss.dbeaver.utils.ContentUtils;
+import org.jkiss.dbeaver.utils.GeneralUtils;
 import org.jkiss.dbeaver.utils.RuntimeUtils;
+import org.jkiss.dbeaver.utils.SystemVariablesResolver;
 import org.jkiss.utils.CommonUtils;
+import org.jkiss.utils.StandardConstants;
 import org.osgi.framework.Bundle;
 
+import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.*;
 
 /**
@@ -67,6 +74,7 @@ public abstract class BasePlatformImpl implements DBPPlatform, DBPApplicationCon
 
     private static final String APP_CONFIG_FILE = "dbeaver.ini";
     private static final String ECLIPSE_CONFIG_FILE = "eclipse.ini";
+    private static final String TEMP_PROJECT_NAME = ".dbeaver-temp"; //$NON-NLS-1$
 
     public static final String CONFIG_FOLDER = ".config";
     public static final String FILES_FOLDER = ".files";
@@ -86,6 +94,8 @@ public abstract class BasePlatformImpl implements DBPPlatform, DBPApplicationCon
     private SQLDialectMetadataRegistry sqlDialectRegistry;
 
     private DBPPlatformLanguage platformLanguage;
+
+    protected Path tempFolder;
 
     protected void initialize() {
         log.debug("Initialize base platform...");
@@ -149,6 +159,13 @@ public abstract class BasePlatformImpl implements DBPPlatform, DBPApplicationCon
         }
         activatedServices.clear();
 
+        // Remove temp folder
+        if (tempFolder != null) {
+            if (!ContentUtils.deleteFileRecursive(tempFolder)) {
+                log.warn("Can not delete temp folder '" + tempFolder + "'");
+            }
+            tempFolder = null;
+        }
         // Dispose navigator model first
         // It is a part of UI
         disposeNavigatorModel();
@@ -378,6 +395,64 @@ public abstract class BasePlatformImpl implements DBPPlatform, DBPApplicationCon
         // This property is fake. But we set it to trigger property change listener
         // which will ask to restart workbench.
         getPreferenceStore().setValue(ModelPreferences.PLATFORM_LANGUAGE, language.getCode());
+    }
+
+    @NotNull
+    public Path getTempFolder(@NotNull DBRProgressMonitor monitor, @NotNull String name) {
+        if (tempFolder == null) {
+            // Make temp folder
+            try {
+                String tempFolderPath = System.getProperty("dbeaver.io.tmpdir");
+                if (!CommonUtils.isEmpty(tempFolderPath)) {
+                    tempFolderPath = GeneralUtils.replaceVariables(tempFolderPath, new SystemVariablesResolver());
+
+                    File dbTempFolder = new File(tempFolderPath);
+                    if (!dbTempFolder.mkdirs()) {
+                        throw new IOException("Can't create temp directory '" + dbTempFolder.getAbsolutePath() + "'");
+                    }
+                } else {
+                    tempFolderPath = System.getProperty(StandardConstants.ENV_TMP_DIR);
+                }
+                monitor.subTask("Create temp folder '" + tempFolderPath + "'");
+                Path tmpFolder = Paths.get(tempFolderPath);
+                if (!Files.exists(tmpFolder)) {
+                    log.debug("Create global temp folder '" + tmpFolder + "'");
+                    Files.createDirectories(tmpFolder);
+                }
+                tempFolder = Files.createTempDirectory(tmpFolder, TEMP_PROJECT_NAME);
+            } catch (IOException e) {
+                final String sysTempFolder = System.getProperty(StandardConstants.ENV_TMP_DIR);
+                if (!CommonUtils.isEmpty(sysTempFolder)) {
+                    tempFolder = Path.of(sysTempFolder).resolve(TEMP_PROJECT_NAME);
+                    if (!Files.exists(tempFolder)) {
+                        try {
+                            Files.createDirectories(tempFolder);
+                        } catch (IOException ex) {
+                            final String sysUserFolder = System.getProperty(StandardConstants.ENV_USER_HOME);
+                            if (!CommonUtils.isEmpty(sysUserFolder)) {
+                                tempFolder = Path.of(sysUserFolder).resolve(TEMP_PROJECT_NAME);
+                                if (!Files.exists(tempFolder)) {
+                                    try {
+                                        Files.createDirectories(tempFolder);
+                                    } catch (IOException exc) {
+                                        tempFolder = Path.of(TEMP_PROJECT_NAME);
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        }
+        Path localTemp = name == null ? tempFolder : tempFolder.resolve(name);
+        if (!Files.exists(localTemp)) {
+            try {
+                Files.createDirectories(localTemp);
+            } catch (IOException e) {
+                log.error("Can't create temp directory " + localTemp, e);
+            }
+        }
+        return localTemp;
     }
 
 }
